[#211] Payment 테이블 정의

+++ queuing-system/queuing-system-for-java/src/main/java/io/chan/queuingsystemforjava/domain/payment/dto/PaymentRequest.java
@@ -1,8 +1,10 @@
 package io.chan.queuingsystemforjava.domain.payment.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentRequest(
         String paymentKey,
         String orderId,
-        long amount
+        BigDecimal amount
 ) {
 }
