+++ queuing-system/queuing-system-for-java/src/main/java/io/chan/queuingsystemforjava/domain/payment/service/PaymentService.java
@@ -6,6 +6,8 @@ import io.chan.queuingsystemforjava.domain.payment.dto.PaymentResponse;
 import io.github.resilience4j.retry.annotation.Retry;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 @Service
 public class PaymentService {
     private final PaymentApiClient paymentApiClient;
@@ -15,12 +17,12 @@ public class PaymentService {
     }
 
     @Retry(name = "paymentRetry", fallbackMethod = "fallback")
-    public PaymentResponse confirmPayment(String paymentKey, String orderId, long amount) {
+    public PaymentResponse confirmPayment(String paymentKey, String orderId, BigDecimal amount) {
         PaymentRequest request = new PaymentRequest(paymentKey, orderId, amount);
         return paymentApiClient.confirmPayment(request);
     }
 
-    public PaymentResponse fallback(String paymentKey, String orderId, long amount, Throwable t) {
+    public PaymentResponse fallback(String paymentKey, String orderId, BigDecimal amount, Throwable t) {
         if (t instanceof IllegalArgumentException || t instanceof IllegalStateException) {
             throw (RuntimeException) t; // 원래 예외를 그대로 던짐
         }
